Attempt to reduce flakes in window_test.
The tests seem prone to flaking if executed around the hour. This wouldn't necessarily affect real world outcomes, but the test wants the result to be in the previous hour if the tested code actually executes just after the hour rolls over.

Try to reduce the number of places that Now() is used, so that the entire test(s) are referring to the same point in time.

PiperOrigin-RevId: 433552521